+++ changelog.md
@@ -16,7 +16,9 @@ Changes:
 
 * a new avatar added and a new quotation added into php database page
 
-make quotation show on page
+* make quotation show on page
+
+* add random() method